layers: Move SemaphoreSubmitState
This type is used as a local variable by QueueSubmit to collect updates
during validation. It is not a vvl state object that tracks state
globally.
