+++ BUILD.gn
@@ -129,6 +129,7 @@ vvl_sources = [
   "layers/core_checks/cc_submit.cpp",
   "layers/core_checks/cc_submit.h",
   "layers/core_checks/cc_synchronization.cpp",
+  "layers/core_checks/cc_synchronization.h",
   "layers/core_checks/cc_video.cpp",
   "layers/core_checks/cc_vuid_maps.cpp",
   "layers/core_checks/cc_vuid_maps.h",
