+++ layers/core_checks/cc_queue.cpp
@@ -22,6 +22,7 @@
 
 #include <vulkan/vk_enum_string_helper.h>
 #include "generated/chassis.h"
+#include "cc_synchronization.h"
 #include "core_validation.h"
 #include "state_tracker/queue_state.h"
 #include "state_tracker/semaphore_state.h"
